Move cleanup to intervention graph. Make debug true by default
